% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{nmu_assessment}
\alias{nmu_assessment}
\title{NMU Assessment Document}
\usage{
nmu_assessment(fontsize = "12pt", ...)
}
\arguments{
\item{fontsize}{Font size (e.g., "12pt")}

\item{...}{Additional arguments to pdf_document}
}
\description{
NMU Assessment Document
}
