% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{nmu_beamer}
\alias{nmu_beamer}
\title{NMU Beamer Presentation}
\usage{
nmu_beamer(fontsize = "12pt", ...)
}
\arguments{
\item{fontsize}{Font size (e.g., "12pt")}

\item{...}{Additional arguments to beamer_presentation}
}
\description{
NMU Beamer Presentation
}
