% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{nmu_document}
\alias{nmu_document}
\title{NMU General Document}
\usage{
nmu_document(fontsize = "12pt", ...)
}
\arguments{
\item{fontsize}{Font size (e.g., "12pt")}

\item{...}{Additional arguments to pdf_document}
}
\description{
NMU General Document
}
