% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{nmu_assessment}
\alias{nmu_assessment}
\title{NMU Assessment Document}
\usage{
nmu_assessment(solutions = FALSE, ...)
}
\arguments{
\item{solutions}{Include solutions? (logical, default FALSE)}

\item{...}{Additional arguments to pdf_document}
}
\description{
NMU Assessment Document
}
